Extract interface from BuildOperationAncestryTracker
