+++ subprojects/build-operations/src/main/java/org/gradle/internal/operations/BuildOperationAncestryTracker.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,51 +17,13 @@
 package org.gradle.internal.operations;
 
 import javax.annotation.Nullable;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
 @SuppressWarnings("Since15")
-public class BuildOperationAncestryTracker implements BuildOperationListener {
+public interface BuildOperationAncestryTracker {
+    Optional<OperationIdentifier> findClosestMatchingAncestor(@Nullable OperationIdentifier id, Predicate<? super OperationIdentifier> predicate);
 
-    private final Map<OperationIdentifier, OperationIdentifier> parents = new ConcurrentHashMap<OperationIdentifier, OperationIdentifier>();
-
-    public Optional<OperationIdentifier> findClosestMatchingAncestor(@Nullable OperationIdentifier id, Predicate<? super OperationIdentifier> predicate) {
-        if (id == null) {
-            return Optional.empty();
-        }
-        if (predicate.test(id)) {
-            return Optional.of(id);
-        }
-        return findClosestMatchingAncestor(parents.get(id), predicate);
-    }
-
-    public <T> Optional<T> findClosestExistingAncestor(@Nullable OperationIdentifier id, Function<? super OperationIdentifier, T> lookupFunction) {
-        if (id == null) {
-            return Optional.empty();
-        }
-        T value = lookupFunction.apply(id);
-        if (value != null) {
-            return Optional.of(value);
-        }
-        return findClosestExistingAncestor(parents.get(id), lookupFunction);
-    }
-
-    @Override
-    public void started(BuildOperationDescriptor buildOperation, OperationStartEvent startEvent) {
-        if (buildOperation.getParentId() != null) {
-            parents.put(buildOperation.getId(), buildOperation.getParentId());
-        }
-    }
-
-    @Override
-    public void progress(OperationIdentifier operationIdentifier, OperationProgressEvent progressEvent) {
-    }
-
-    @Override
-    public void finished(BuildOperationDescriptor buildOperation, OperationFinishEvent finishEvent) {
-        parents.remove(buildOperation.getId());
-    }
+    <T> Optional<T> findClosestExistingAncestor(@Nullable OperationIdentifier id, Function<? super OperationIdentifier, T> lookupFunction);
 }
